fix(entries): rename entries
