+++ src/components/Entries/Entries.queries.tsx
