+++ src/components/Entries/EntriesList/styled.tsx
