+++ src/components/Entries/NewEntryForm/NewEntryForm.component.tsx
