+++ src/components/Entries/NewEntryForm/styled.tsx
