+++ src/components/index.ts
@@ -1,4 +1,4 @@
-export * from './EntriesTest';
+export * from './Entries';
 export * from './RegisterForm';
 export * from './LoginForm';
 export * from './Logout';
